Add certificate revocation checking, off by default (RhBug:1814383)
This adds libcurl option CURLOPT_SSL_VERIFYSTATUS which is now disabled by
defualt. The options is also exposed in the API (C and Python) and there
are some python tests for it.

Note that if this option is enabled but the server does not support the
TLS extension (OCSP sapling), the verification will fail.

https://bugzilla.redhat.com/show_bug.cgi?id=1814383